after many changes on appointments

+++ app/Http/Controllers/appointmentController.php
@@ -10,47 +10,57 @@ use Illuminate\Support\Facades\Auth;
 
 class appointmentController extends Controller
 {
-    public function __construct()
+/*    public function __construct()
     {
         $this->middleware('auth:api');
-    }
+    }*/
     public function sched_appointment(Request $request)
     {
-        $userid = $_GET['userid'];
-        $doctor = Doctor::where('user_id', $userid)->first();
-        $doctor_id = $doctor->id;
+        if(!Auth::user()){
+            $state="not authorized to access";
+            $message="cannot access to api resources";
+            $data = [
+                'isUser'=>0
+            ];
+            return response(compact('state', 'message','data'),401);
+        }
+        $doctor_id =Auth::user()->id;
+
         $all_data = ($request->input('data'));
         $date = $all_data['date'];
-        $total_slots = $all_data['totalSlots'];
+        $addedAppointments = $all_data['addedAppointments'];
         $date_delete = Appointment::where([['schedule_from', '=', $doctor_id], ['schedule_date', '=', $date], ['appointment_state', '=', 'free']]);
         $date_delete->delete();
-        foreach ($total_slots as $s) {
+        foreach ($addedAppointments as $s) {
             $user = Appointment::create([
                 'schedule_date' => $date,
                 'schedule_from' => $doctor_id,
-                'slot_time' => $s,
+                'slot_time' => $s['slotTime'],
+                'appointment_type'=> $s['appointmentType'],
+                'duration'=>$s['appointmentDuration'],
             ]);
         }
-        $a=Auth::user()->id;
         $msg = "done";
-        return response(compact('msg','a'), 200);
-
-
+        return response(compact('msg','doctor_id'), 200);
     }
 
     /*
      * ******************** Book Appointment ********************
      */
     public function book_appointment(Request $request){
-        $userid = $_GET['userid'];
-        $user = User::where('id', $userid)->first();
-        $user_id = $user->id;
+        if(!Auth::user()){
+            $state="not authorized to access";
+            $message="cannot access to api resources";
+            $data = [
+                'isUser'=>0
+            ];
+            return response(compact('state', 'message','data'),401);
+        }
+        $user_id =Auth::user()->id;
         $all_data = ($request->input('data'));
         $date = $all_data['date'];
         $booked_slot = $all_data['bookedSlot'];
-        $doc_user = $all_data['doctorId'];
-        $doctor = Doctor::where('user_id', $doc_user)->first();
-        $doctor_id = $doctor->id;
+        $doctor_id = $all_data['doctorId'];
         $date_update = Appointment::where([['schedule_from', '=', $doctor_id], ['schedule_date', '=', $date], ['slot_time', '=', $booked_slot]])->first();
 
         if(empty($date_update)){
@@ -90,33 +100,44 @@ class appointmentController extends Controller
 
     }
     /*
-     *
-     **/
+    * ******************** get_slots ********************
+    */
     public function get_slots(Request $request){
 
         $all_data = ($request->input('data'));
         $date = $all_data['date'];
-        $doc_user = $all_data['doctorId'];
-        $doctor = Doctor::where('user_id', $doc_user)->first();
-        $doctor_id = $doctor->id;
+        $doctor_id = $all_data['doctorId'];
         $totalSlots=array();
         $bookedSlots=array();
-        $freeSlots=array();;
+        $freeSlots=array();
         $ts = Appointment::where([['schedule_from', '=', $doctor_id], ['schedule_date', '=', $date]])->get();
 /*        return response(compact('ts'),200);*/
+        $slot=array('appointmentState'=>'0','appointmentType'=>'none','slotTime'=>'12:00 am','appointmentDuration'=>'0' );
 
         $bs = Appointment::where([['schedule_from', '=', $doctor_id], ['schedule_date', '=', $date], ['appointment_state', '=', 'booked']])->get();
         $fs = Appointment::where([['schedule_from', '=', $doctor_id], ['schedule_date', '=', $date], ['appointment_state', '=', 'free']])->get();
         foreach ($ts as $t_s){
-            array_push($totalSlots, $t_s->slot_time);
+            $slot['appointmentState']=$t_s->appointment_state;
+            $slot['appointmentType']=$t_s->appointment_type;
+            $slot['slotTime']=$t_s->slot_time;
+            $slot['appointmentDuration']=$t_s->duration;
+            array_push($totalSlots, $slot);
         }
 
         foreach ($bs as $b_s){
-            array_push($bookedSlots, $b_s->slot_time);
+            $slot['appointmentState']=$b_s->appointment_state;
+            $slot['appointmentType']=$b_s->appointment_type;
+            $slot['slotTime']=$b_s->slot_time;
+            $slot['appointmentDuration']=$b_s->duration;
+            array_push($bookedSlots, $slot);
         }
-        $freeSlots=array();
+
         foreach ($fs as $f_s) {
-            array_push($freeSlots, $f_s->slot_time);
+            $slot['appointmentState']=$f_s->appointment_state;
+            $slot['appointmentType']=$f_s->appointment_type;
+            $slot['slotTime']=$f_s->slot_time;
+            $slot['appointmentDuration']=$f_s->duration;
+            array_push($freeSlots, $slot);
         }
         if (empty($freeSlots)) {
             $freeSlots=null;
@@ -130,6 +151,7 @@ class appointmentController extends Controller
         ];
         return response(compact('state', 'message','data'),200);
     }
+
     public function cancel_appointment(Request $request){
         $all_data = ($request->input('data'));
         $date = $all_data['date'];
@@ -138,15 +160,12 @@ class appointmentController extends Controller
         $state="state";
         $message="information retreived successfully";
         if($cancelFrom=='doctor'){
-            $doc_user = $_GET['userid'];;
-            $doctor = Doctor::where('user_id', $doc_user)->first();
-            $doctor_id = $doctor->id;
+            $doctor_id = $_GET['userid'];
             $data_update = Appointment::where([['schedule_from', '=', $doctor_id], ['schedule_date', '=', $date], ['slot_time', '=', $bookedSlot]])->first();
-
             $data_update->appointment_state="cancelled";
             $data_update->save();
             $data = [
-                'fromDoctor'=>$doc_user,
+                'fromDoctor'=>$doctor_id,
             ];
             return response(compact('state', 'message','data'),200);
